+++ src/MeshIO.FBX.Tests/FbxReaderTest.cs
@@ -29,12 +29,12 @@ namespace MeshIO.FBX.Tests
 
 		public FbxReaderTest(ITestOutputHelper output) : base(output) { }
 
-		[Theory]
+		[Theory(Skip = "skipy skip")]
 		[MemberData(nameof(AsciiFiles))]
 		[MemberData(nameof(BinaryFiles))]
 		public void GetVersion(string path)
 		{
-			using (FbxReader reader = new FbxReader(path, ErrorLevel.Checked))
+			using (FbxReader reader = new FbxReader(path))
 			{
 				var version = reader.GetVersion();
 			}
@@ -75,22 +75,9 @@ namespace MeshIO.FBX.Tests
 
 		private Scene readFile(string path)
 		{
-			using (FbxReader reader = new FbxReader(path, ErrorLevel.Checked))
+			using (FbxReader reader = new FbxReader(path))
 			{
 				reader.OnNotification += onNotification;
-
-				if (reader.GetVersion() <= FbxVersion.v5800)
-				{
-					Assert.Throws<NotSupportedException>(reader.Read);
-					return null;
-				}
-
-				if (reader.GetVersion() <= FbxVersion.v6100)
-				{
-					Assert.Throws<NotImplementedException>(reader.Read);
-					return null;
-				}
-
 				return reader.Read();
 			}
 		}
