--- src/MeshIO.FBX/Converters/ISceneConverter.cs
@@ -1,12 +1,0 @@
-﻿namespace MeshIO.FBX.Converters
-{
-	/// <summary>
-	/// Converts a fbx scene to a <see cref="FbxRootNode"/>
-	/// </summary>
-	public interface ISceneConverter
-	{
-		FbxVersion Version { get; }
-
-		FbxRootNode ToRootNode();
-	}
-}
