--- src/MeshIO.FBX/Converters/SceneConverter7000.cs
@@ -1,8 +1,0 @@
-﻿namespace MeshIO.FBX.Converters
-{
-	/// <inheritdoc/>
-	public class SceneConverter7000 : SceneConverterBase
-	{
-		public SceneConverter7000(Scene scene) : base(scene, FbxVersion.v7400) { }
-	}
-}
