+++ src/MeshIO.FBX/FbxNodeCollection.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections;
+﻿using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -8,7 +7,6 @@ namespace MeshIO.FBX
 	/// <summary>
 	/// Base class for nodes and documents
 	/// </summary>
-	[Obsolete("This class will be removed")]
 	public abstract class FbxNodeCollection : IEnumerable<FbxNode>
 	{
 		/// <summary>
@@ -71,7 +69,6 @@ namespace MeshIO.FBX
 		public bool TryGetNode(string name, out FbxNode node)
 		{
 			node = this[name];
-
 			return node != null;
 		}
