+++ src/MeshIO.FBX/FbxReaderOptions.cs
@@ -1,0 +1,7 @@
+﻿namespace MeshIO.FBX
+{
+	public class FbxReaderOptions
+	{
+		public ErrorLevel ErrorLevel { get; set; } = ErrorLevel.Permissive;
+	}
+}
