--- src/MeshIO.FBX/Parsers/IFbxParser.cs
@@ -1,7 +1,0 @@
-﻿namespace MeshIO.FBX
-{
-	internal interface IFbxParser
-	{
-		FbxRootNode Parse();
-	}
-}
