+++ src/MeshIO.FBX/Readers/FbxFileReader7000.cs
@@ -1,0 +1,7 @@
+﻿namespace MeshIO.FBX.Readers
+{
+	internal class FbxFileReader7000 : FbxFileReaderBase
+	{
+		public FbxFileReader7000(FbxRootNode root, FbxReaderOptions options) : base(root, options) { }
+	}
+}
