+++ src/MeshIO.FBX/Readers/Parsers/IFbxParser.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace MeshIO.FBX.Readers.Parsers
+{
+	internal interface IFbxParser : IDisposable
+	{
+		FbxRootNode Parse();
+	}
+}
