+++ src/MeshIO.FBX/Readers/Templates/IFbxObjectTemplate.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MeshIO.FBX.Readers.Templates
+{
+	internal interface IFbxObjectTemplate
+	{
+	}
+}
